HW1

+++ src/main/java/org/example/HomeWorks/HomeWork1.java
@@ -6,7 +6,7 @@ import java.util.OptionalDouble;
 public class HomeWork1 {
     public static double average(List<Double> list) {
         return list.stream()
-                .filter(n -> n.doubleValue() % 2 == 0)
+                .filter(n -> n % 2 == 0)
                 .mapToDouble(Double::doubleValue)
                 .average()
                 .orElse(0.0);
